avoid circular deps
